COMMON: Fix MSVC warning in PEResources::parseResourceLevel()

+++ common/winexe_pe.cpp
@@ -123,7 +123,7 @@ void PEResources::parseResourceLevel(Section &section, uint32 offset, int level)
 	uint16 namedEntryCount = _exe->readUint16LE();
 	uint16 intEntryCount = _exe->readUint16LE();
 
-	for (uint32 i = 0; i < namedEntryCount + intEntryCount; i++) {
+	for (uint32 i = 0; i < (uint32)(namedEntryCount + intEntryCount); i++) {
 		uint32 value = _exe->readUint32LE();
 
 		WinResourceID id;
